+++ src/assets/audio/index.ts
@@ -1,0 +1,3 @@
+import toothlessDancingSong from "./toothless-dancing-song.mp4";
+
+export { toothlessDancingSong };
