+++ src/assets/image/index.ts
@@ -1,0 +1,9 @@
+import portfolioImg1 from "./portfolio_img1.jpg";
+import portfolioImg2 from "./portfolio_img2.jpg";
+import portfolioImg3 from "./portfolio_img3.jpg";
+import postThumbNailSample from "./post_thumb_nail_sample.png";
+import profileImage from "./profile-image.png";
+import riaxoLogo from "./riaxo-logo.png";
+import toothlessDancing from "./toothless-dancing.gif";
+
+export { portfolioImg1, portfolioImg2, portfolioImg3, postThumbNailSample, profileImage, riaxoLogo, toothlessDancing };
