+++ src/components/@shared/PostTag/PostTag.tsx
@@ -1,5 +1,6 @@
 import { T } from "../../../styles/TextGuide.styles";
-import * as S from "./index.styles";
+import * as S from "./PostTag.styles";
+
 export interface IPostTag {
     name?: string;
     size?: "large" | "small";
