+++ src/index.tsx
@@ -1,13 +1,16 @@
-import ReactDOM from "react-dom/client";
-import App from "./App";
 import "./styles/colorRoot.css";
-import { BrowserRouter } from "react-router-dom";
-import { GlobalStyle } from "./styles/globalStyles";
-import { RecoilRoot } from "recoil";
-import { Layout } from "./layout/Layout";
-import { globalTheme } from "./styles/theme.styles";
-import { ThemeProvider } from "@mui/material";
+
+import ReactDOM from "react-dom/client";
 import { QueryClient, QueryClientProvider } from "react-query";
+import { BrowserRouter } from "react-router-dom";
+import { RecoilRoot } from "recoil";
+
+import { ThemeProvider } from "@mui/material";
+
+import App from "./App";
+import { Layout } from "./layout/Layout";
+import { GlobalStyle } from "./styles/globalStyles";
+import { globalTheme } from "./styles/theme.styles";
 
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
 const queryClient = new QueryClient();
