+++ src/pages/Main/Main.tsx
@@ -1,7 +1,5 @@
+import { IntroBox, MainPagePostCardList, MainPagePostList } from "./components/index";
 import * as S from "./Main.styles";
-import { IntroBox } from "./components/IntroBox/IntroBox";
-import { MainPagePostCardList } from "./components/MainPagePostCardList/MainPagePostCardList";
-import { MainPagePostList } from "./components/MainPagePostList/MainPagePostList";
 
 export const Main: React.FC = () => {
     return (
