+++ src/pages/Main/components/MainPagePostList/MainPagePostList.tsx
@@ -1,10 +1,9 @@
+import { fetchPostList } from "../../../../api/index";
+import { useCustomQuery } from "../../../../hooks/index";
 import { T } from "../../../../styles/TextGuide.styles";
+import { IPost } from "../../../../types/index";
+import { formatDateFromAPIToYYYYMMDD, getQueryString } from "../../../../utils/index";
 import * as S from "./MainPagePostList.styles";
-import { IPost } from "../../../../types/post";
-import { getQueryString } from "../../../../utils/getQueryString";
-import { fetchPostList } from "../../../../api/post.queries";
-import { formatDateFromAPIToYYYYMMDD } from "../../../../utils/formatDate";
-import { useCustomQuery } from "../../../../hooks/useCustomQuery";
 
 export const MainPagePostList: React.FC = () => {
     const queryString = getQueryString();
