+++ src/pages/Main/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./IntroBox/IntroBox";
+export * from "./MainPagePostCardList/MainPagePostCardList";
+export * from "./MainPagePostList/MainPagePostList";
