+++ src/pages/Post/components/PrevNextNavBox/PrevNextNavBox.tsx
@@ -1,8 +1,9 @@
-import { T } from "../../../../styles/TextGuide.styles";
-import { IPost } from "../../../../types/post";
-import * as S from "./PrevNextNavBox.styles";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
 import ArrowForwardIcon from "@mui/icons-material/ArrowForward";
+
+import { T } from "../../../../styles/TextGuide.styles";
+import { IPost } from "../../../../types/index";
+import * as S from "./PrevNextNavBox.styles";
 
 interface IPrevNextNavBox {
     postData?: IPost[];
