+++ src/pages/Post/components/SideNavBox/SideNavBox.tsx
@@ -1,4 +1,5 @@
 import { useEffect, useState } from "react";
+
 import { T } from "../../../../styles/TextGuide.styles";
 import * as S from "./SideNavBox.styles";
