+++ src/pages/Post/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./PrevNextNavBox/PrevNextNavBox";
+export * from "./SideNavBox/SideNavBox";
