+++ src/pages/PostCreate/PostCreate.tsx
@@ -1,23 +1,21 @@
-import * as S from "./PostCreate.styles";
-import { Autocomplete, Button, Chip, Skeleton, TextField } from "@mui/material";
 import { useEffect, useRef, useState } from "react";
-import { Editor } from "@tinymce/tinymce-react";
-import { BASE_URL, TINY_MCE_API_KEY } from "../../constants/API";
-import { EDIT_TOOLBAR, PLUGINS } from "../../constants/tinyMceOption";
-import { createPost, fetchPostList, updatePost } from "../../api/post.queries";
-import { useLocation, useNavigate } from "react-router-dom";
-import { IPost } from "../../types/post";
-import { getQueryString } from "../../utils/getQueryString";
-import { ProfileImgUpload } from "./components/ImageUpload/ImageUpload";
-import { createImageUpload } from "../../api/resource.queries";
-import { useRecoilState } from "recoil";
-import { isLoadingState } from "../../recoil/atoms";
-import { useSnackbar } from "notistack";
-import { ImageGenerateBox } from "./components/ImageGenerateBox/ImageGenerateBox";
-import { fetchTagList } from "../../api/tag.queries";
-import { useCustomQuery } from "../../hooks/useCustomQuery";
-import { useCustomMutation } from "../../hooks/useCustomMutation";
+
 import axios from "axios";
+import { useSnackbar } from "notistack";
+import { useLocation, useNavigate } from "react-router-dom";
+import { useRecoilState } from "recoil";
+
+import { Autocomplete, Button, Chip, Skeleton, TextField } from "@mui/material";
+import { Editor } from "@tinymce/tinymce-react";
+
+import { createImageUpload, createPost, fetchPostList, fetchTagList, updatePost } from "../../api/index";
+import { BASE_URL, EDIT_TOOLBAR, PLUGINS, TINY_MCE_API_KEY } from "../../constants/index";
+import { useCustomMutation, useCustomQuery } from "../../hooks/index";
+import { isLoadingState } from "../../recoil/atoms/index";
+import { IPost } from "../../types/index";
+import { getQueryString } from "../../utils/index";
+import { ImageGenerateBox, ProfileImgUpload } from "./components/index";
+import * as S from "./PostCreate.styles";
 
 export const PostCreate: React.FC = () => {
     const editorRef = useRef<any>(null);
