+++ src/pages/PostCreate/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./ImageGenerateBox/ImageGenerateBox";
+export * from "./ImageUpload/ImageUpload";
